# MovieBattleOpenAPIDefinition.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/quizzes/questions/answer": {
      "post": {
        "tags": [
          "quiz-controller"
        ],
        "operationId": "answerQuestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Response"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "selectedOption": {
            "$ref": "#/components/schemas/MovieCard"
          },
          "correct": {
            "type": "boolean"
          }
        }
      },
      "MovieCard": {
        "type": "object",
        "properties": {
          "imdbId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "releasedYear": {
            "type": "string"
          },
          "posterUrl": {
            "type": "string"
          },
          "imdbRating": {
            "type": "number",
            "format": "double"
          },
          "imdbVotes": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "imdbId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oAuth2CommonFlow": {
        "type": "oauth2",
        "description":"This API uses OAuth2 with the password grant type flow.",
        "flows": [
          {
            "password": {
              "tokenUrl": "http://localhost:8083/oauth/token",
              "refreshUrl": "http://localhost:8083/oauth/token",
              "scopes": [
                {
                  "play": ""
                }
              ]
            }
          }
        ]
      }
    }
  }
}